requirements.txt
control==0.10.1
decorator==5.1.1
matplotlib==3.9.2
numpy==1.26.4
pandas==2.2.2
pillow==10.4.0
plotly==5.24.1
scipy==1.14.1
pyserial==3.5
streamlit==1.40.2
tornado==6.4.2